Bugfixes & improvements
- double instead of float
- Enable greedy heuristics (sorting points)
- Create max_arch() to benchmark against stochastic implementation
- Fix bugs related to comparison operators
- Enable new options
- Set cout precision

+++ head/temperature.h
@@ -21,12 +21,12 @@
 
 class Temperature {
 private:
-	float curr;
+	double curr;
 public:
-	Temperature(float _curr)
+	Temperature(double _curr)
 		: curr(_curr)
 	{}
-	float operator() (void);
+	double operator() (void);
 };
 
 #endif
